requirements.txt
Django==2.2.0
djangorestframework==3.12.1
django-rest-swagger==2.2.0
django-environ==0.4.4
gunicorn==20.0.4
requests==2.24.0